Debut test service + ptites modifs

+++ src/main/java/fac/luminy/m2/aa1/tp1/repository/VoitureRepository.java
@@ -13,7 +13,7 @@ public interface VoitureRepository extends JpaRepository<Voiture, Long> {
     List<Voiture> findByProprietaireNom(@Param("nom") String nom);
 
     @Query("SELECT v from VOITURE v where ?1 < v.prix * 1.10 or ?1 > v.prix * 0.90")
-    List<Voiture> findByPricePlusMinusTenPercent(@Param("prix") int prix);
+    List<Voiture> findByPricePlusMinusTenPercent(@Param("prix") double prix);
 
     List<Voiture> findVoituresByTypeIn(List<TypeVoiture> typesVoiture);
 }
